Refactor msgserver serial plugin to make it easier for others to create more serial plugins.
